Rename PlayerInputModel to InputRequestModel
