+++ src/client/components/SelectDelegate.vue
@@ -15,7 +15,7 @@
 import Vue from 'vue';
 import AppButton from '@/client/components/common/AppButton.vue';
 import {ColorWithNeutral} from '@/common/Color';
-import {SelectDelegateModel} from '@/common/models/PlayerInputModel';
+import {SelectDelegateModel} from '@/common/models/InputRequestModel';
 import {PublicPlayerModel} from '@/common/models/PlayerModel';
 import SelectPlayerRow from '@/client/components/SelectPlayerRow.vue';
 import {SelectDelegateResponse} from '@/common/inputs/InputResponse';
