+++ src/server/inputs/OrOptions.ts
@@ -1,7 +1,7 @@
 import {InputRequest} from '../InputRequest';
 import {InputResponse, isOrOptionsResponse} from '../../common/inputs/InputResponse';
 import {IPlayer} from '../IPlayer';
-import {OrOptionsModel} from '../../common/models/PlayerInputModel';
+import {OrOptionsModel} from '../../common/models/InputRequestModel';
 import {OptionsInput} from './OptionsPlayerInput';
 
 export class OrOptions extends OptionsInput<undefined> {
