+++ src/server/inputs/SelectOption.ts
@@ -2,7 +2,7 @@ import {Message} from '../../common/logs/Message';
 import {InputRequest} from '../InputRequest';
 import {BaseInputRequest} from '../InputRequest';
 import {InputResponse, isSelectOptionResponse} from '../../common/inputs/InputResponse';
-import {SelectOptionModel} from '../../common/models/PlayerInputModel';
+import {SelectOptionModel} from '../../common/models/InputRequestModel';
 
 export class SelectOption extends BaseInputRequest<undefined> {
   constructor(
