+++ src/server/inputs/SelectPayment.ts
@@ -3,7 +3,7 @@ import {BaseInputRequest} from '../InputRequest';
 import {isPayment, Payment, PaymentOptions} from '../../common/inputs/Payment';
 import {InputResponse, isSelectPaymentResponse} from '../../common/inputs/InputResponse';
 import {IPlayer} from '../IPlayer';
-import {SelectPaymentModel} from '../../common/models/PlayerInputModel';
+import {SelectPaymentModel} from '../../common/models/InputRequestModel';
 
 export class SelectPayment extends BaseInputRequest<Payment> {
   constructor(
