+++ src/server/inputs/SelectSpace.ts
@@ -1,7 +1,7 @@
 import {Message} from '../../common/logs/Message';
 import {Space} from '../boards/Space';
 import {InputResponse, isSelectSpaceResponse} from '../../common/inputs/InputResponse';
-import {SelectSpaceModel} from '../../common/models/PlayerInputModel';
+import {SelectSpaceModel} from '../../common/models/InputRequestModel';
 import {BaseInputRequest} from '../InputRequest';
 
 export class SelectSpace extends BaseInputRequest<Space> {
